+++ kubernetes-client-api/src/main/java/io/fabric8/kubernetes/client/dsl/CertificatesAPIGroupDSL.java
@@ -21,4 +21,6 @@ public interface CertificatesAPIGroupDSL extends Client {
   V1CertificatesAPIGroupDSL v1();
 
   V1beta1CertificatesAPIGroupDSL v1beta1();
+
+  V1Alpha1CertificatesAPIGroupDSL v1alpha1();
 }