+++ kubernetes-client-api/src/main/java/io/fabric8/kubernetes/client/dsl/DynamicResourceAllocationAPIGroupDSL.java
@@ -16,8 +16,8 @@
 package io.fabric8.kubernetes.client.dsl;
 
 import io.fabric8.kubernetes.client.Client;
-import io.fabric8.kubernetes.client.V1Alpha1DynamicResourceAllocationAPIGroupDSL;
+import io.fabric8.kubernetes.client.V1Alpha2DynamicResourceAllocationAPIGroupDSL;
 
 public interface DynamicResourceAllocationAPIGroupDSL extends Client {
-  V1Alpha1DynamicResourceAllocationAPIGroupDSL v1alpha1();
+  V1Alpha2DynamicResourceAllocationAPIGroupDSL v1alpha2();
 }