+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/impl/AuthenticationAPIGroupClient.java
@@ -18,6 +18,7 @@ package io.fabric8.kubernetes.client.impl;
 import io.fabric8.kubernetes.client.V1AuthenticationAPIGroupDSL;
 import io.fabric8.kubernetes.client.dsl.AuthenticationAPIGroupDSL;
 import io.fabric8.kubernetes.client.dsl.V1Alpha1AuthenticationAPIGroupDSL;
+import io.fabric8.kubernetes.client.dsl.V1Beta1AuthenticationAPIGroupDSL;
 import io.fabric8.kubernetes.client.extension.ClientAdapter;
 
 public class AuthenticationAPIGroupClient extends ClientAdapter<AuthenticationAPIGroupClient>
@@ -29,7 +30,12 @@ public class AuthenticationAPIGroupClient extends ClientAdapter<AuthenticationAP
 
   @Override
   public V1Alpha1AuthenticationAPIGroupDSL v1alpha1() {
-    return adapt(V1Alpha1AuthenticationAPIGroupDSL.class);
+    return adapt(V1Alpha1AuthenticationAPIGroupClient.class);
+  }
+
+  @Override
+  public V1Beta1AuthenticationAPIGroupDSL v1beta1() {
+    return adapt(V1Beta1AuthenticationAPIGroupClient.class);
   }
 
   @Override