+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/impl/NetworkAPIGroupClient.java
@@ -19,6 +19,7 @@ import io.fabric8.kubernetes.api.model.networking.v1.NetworkPolicy;
 import io.fabric8.kubernetes.api.model.networking.v1.NetworkPolicyList;
 import io.fabric8.kubernetes.api.model.networking.v1beta1.Ingress;
 import io.fabric8.kubernetes.api.model.networking.v1beta1.IngressList;
+import io.fabric8.kubernetes.client.V1Alpha1NetworkAPIGroupDSL;
 import io.fabric8.kubernetes.client.V1NetworkAPIGroupDSL;
 import io.fabric8.kubernetes.client.V1beta1NetworkAPIGroupDSL;
 import io.fabric8.kubernetes.client.dsl.MixedOperation;
@@ -36,6 +37,11 @@ public class NetworkAPIGroupClient extends ClientAdapter<NetworkAPIGroupClient> 
   @Override
   public V1beta1NetworkAPIGroupDSL v1beta1() {
     return adapt(V1beta1NetworkAPIGroupClient.class);
+  }
+
+  @Override
+  public V1Alpha1NetworkAPIGroupDSL v1alpha1() {
+    return adapt(V1Alpha1NetworkAPIGroupClient.class);
   }
 
   @Override